requirements.txt
asgiref==3.2.3
Django==3.0.2
django-progressbarupload==0.1.7
djangorestframework==3.11.0
iso8601==0.1.12
m3u8==0.5.4
python-ffmpeg-video-streaming==0.0.21
pytz==2019.3
sqlparse==0.3.0
